Extract Buildbot API calls into separate module
